Fix metrics hook PR post comments
