+++ app/src/main/java/com/mapbox/services/android/navigation/testapp/activity/MockNavigationActivity.java
@@ -32,9 +32,9 @@ import com.mapbox.mapboxsdk.maps.MapView;
 import com.mapbox.mapboxsdk.maps.MapboxMap;
 import com.mapbox.mapboxsdk.maps.OnMapReadyCallback;
 import com.mapbox.mapboxsdk.maps.Style;
-import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.MapboxMetricsReporter;
-import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.MetricEvent;
-import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.MetricsObserver;
+import com.mapbox.services.android.navigation.v5.navigation.metrics.MapboxMetricsReporter;
+import com.mapbox.services.android.navigation.v5.navigation.metrics.MetricEvent;
+import com.mapbox.services.android.navigation.v5.navigation.metrics.MetricsObserver;
 import com.mapbox.services.android.navigation.testapp.R;
 import com.mapbox.services.android.navigation.testapp.Utils;
 import com.mapbox.services.android.navigation.testapp.activity.notification.CustomNavigationNotification;
