+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/internal/navigation/RouteRetrievalEvent.kt
@@ -4,8 +4,8 @@ import android.annotation.SuppressLint
 import android.os.Parcelable
 import com.google.gson.Gson
 import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.DirectionsMetrics
-import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.MetricEvent
 import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.NavigationPerformanceEvent
+import com.mapbox.services.android.navigation.v5.navigation.metrics.MetricEvent
 
 @SuppressLint("ParcelCreator")
 internal data class RouteRetrievalEvent(
