--- libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/internal/navigation/metrics/MetricEvent.kt
@@ -1,24 +1,0 @@
-package com.mapbox.services.android.navigation.v5.internal.navigation.metrics
-
-import androidx.annotation.StringDef
-import com.google.gson.Gson
-
-interface MetricEvent {
-
-    @StringDef(
-        DirectionsMetrics.ROUTE_RETRIEVAL,
-        NavigationMetrics.ARRIVE,
-        NavigationMetrics.CANCEL_SESSION,
-        NavigationMetrics.DEPART,
-        NavigationMetrics.REROUTE,
-        NavigationMetrics.FEEDBACK,
-        NavigationMetrics.INITIAL_GPS,
-        NavigationMetrics.APP_USER_TURNSTILE
-    )
-    annotation class Metric
-
-    @Metric
-    val metric: String
-
-    fun toJson(gson: Gson): String
-}
