--- libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/internal/navigation/metrics/MetricsObserver.kt
@@ -1,6 +1,0 @@
-package com.mapbox.services.android.navigation.v5.internal.navigation.metrics
-
-interface MetricsObserver {
-
-    fun onMetricUpdated(@MetricEvent.Metric metric: String, jsonStringData: String)
-}
