--- libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/internal/navigation/metrics/MetricsReporter.kt
@@ -1,8 +1,0 @@
-package com.mapbox.services.android.navigation.v5.internal.navigation.metrics
-
-interface MetricsReporter {
-
-    fun addEvent(metricEvent: MetricEvent)
-
-    fun setMetricsObserver(metricsObserver: MetricsObserver)
-}
