+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/internal/utils/extensions/GsonEx.kt
@@ -1,0 +1,8 @@
+@file:JvmName("GsonEx")
+
+package com.mapbox.services.android.navigation.v5.internal.utils.extensions
+
+import com.google.gson.Gson
+
+internal inline fun <reified T> Gson.fromJson(string: String): T =
+    this.fromJson(string, T::class.java)
