--- libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/utils/extensions/GsonEx.kt
@@ -1,8 +1,0 @@
-@file:JvmName("GsonEx")
-
-package com.mapbox.services.android.navigation.v5.utils.extensions
-
-import com.google.gson.Gson
-
-inline fun <reified T> Gson.fromJson(string: String): T =
-    this.fromJson(string, T::class.java)
