--- libnavigation-base/src/main/java/com/mapbox/navigation/base/metrics/DirectionsMetrics.kt
@@ -1,8 +1,0 @@
-package com.mapbox.navigation.base.metrics
-
-interface DirectionsMetrics {
-
-    companion object {
-        const val ROUTE_RETRIEVAL = "route_retrieval_event"
-    }
-}
