+++ libnavigation-base/src/main/java/com/mapbox/navigation/base/metrics/MetricEvent.kt
@@ -2,6 +2,8 @@ package com.mapbox.navigation.base.metrics
 
 import androidx.annotation.StringDef
 import com.google.gson.Gson
+import com.mapbox.navigation.base.internal.metrics.DirectionsMetrics
+import com.mapbox.navigation.base.internal.metrics.NavigationMetrics
 
 interface MetricEvent {
