+++ libnavigation-metrics/src/test/java/com/mapbox/navigation/metrics/MapboxMetricsReporterTest.kt
@@ -5,7 +5,6 @@ import com.google.gson.Gson
 import com.mapbox.android.telemetry.Event
 import com.mapbox.android.telemetry.MapboxTelemetry
 import com.mapbox.navigation.base.metrics.MetricEvent
-import com.mapbox.navigation.base.metrics.NavigationMetrics
 import com.mapbox.navigation.metrics.extensions.toTelemetryEvent
 import com.mapbox.navigation.utils.thread.WorkThreadHandler
 import io.mockk.mockk
@@ -36,7 +35,7 @@ class MapboxMetricsReporterTest {
     @Test
     fun telemetryPushCalledWhenAddValidEvent() {
         val mapboxTelemetry = initMetricsReporterWithTelemetry()
-        val metricEvent = StubNavigationEvent(NavigationMetrics.ARRIVE)
+        val metricEvent = StubNavigationEvent("navigation.arrive")
         val event = metricEvent.toTelemetryEvent()
 
         MapboxMetricsReporter.addEvent(metricEvent)
